+++ crates/ff_kinetics/src/macrostates.rs
@@ -330,7 +330,7 @@ impl<'a, E: EnergyModel> MacrostateRegistry<'a, E> {
 mod tests {
     use super::*;
     use std::io::Cursor;
-    use ff_energy::ViennaRNA;
+    use ff_energy::{parameters::RNA_TURNER_2004, ViennaRNA};
 
     #[test]
     fn test_macrostate_init() {
@@ -347,7 +347,7 @@ mod tests {
         .(((.(...)))).((((........))))...............
         */        
 
-        let energy_model = ViennaRNA::default();
+        let energy_model = ViennaRNA::from_thermo_params(&RNA_TURNER_2004, 37.0);
         let seq = NucleotideVec::try_from("UCAGUCUUCGCUGCGCUGUAUCGAUUCGGUUUCAGUUUUUAUUGC").unwrap();
         let db1 = DotBracketVec::try_from(".((((....)))).((((........))))...............").unwrap();
         let db2 = DotBracketVec::try_from(".((((....)))).((((.(....).))))...............").unwrap();
@@ -381,7 +381,7 @@ mod tests {
 
     #[test]
     fn test_macrostateregistry_init_and_classify() {
-        let energy_model = ViennaRNA::default();
+        let energy_model = ViennaRNA::from_thermo_params(&RNA_TURNER_2004, 37.0);
         let seq = NucleotideVec::try_from("UCAGUCUUCGCUGCGCUGUAUCGAUUCGGUUUCAGUUUUUAUUGC").unwrap();
 
         let mut registry = MacrostateRegistry::from((&seq, &energy_model));