+++ crates/ff_python/src/energy_exports.rs
@@ -5,27 +5,14 @@ use ff_structure::MultiPairTable;
 use ff_energy::NucleotideVec;
 use ff_energy::EnergyModel;
 use ff_energy::ViennaRNA as VRNA;
+use ff_energy::parameters::RNA_EXTENDED;
 use ff_energy::parameters::RNA_TURNER_2004;
-
-#[pyfunction]
-pub fn eval(sequence: &str, structure: &str) -> PyResult<i32> {
-    let model = VRNA::default();
-
-    let sequence = NucleotideVec::try_from(sequence)
-        .map_err(|e| PyValueError::new_err(format!("Invalid sequence: {:?}", e)))?;
-
-    let pairings = MultiPairTable::try_from(structure)
-        .map_err(|e| PyValueError::new_err(format!("Invalid structure: {:?}", e)))?;
-
-    let energy = model.energy_of_structure(&sequence, &pairings)
-        .map_err(|e| PyValueError::new_err(format!("Energy evaluation errored: {:?}", e)))?;
-
-    Ok(energy)
-}
+use ff_energy::parameters::DNA_MATHEWS_2004;
 
 #[pyclass]
 pub struct ViennaRNA {
     model: VRNA,
+    is_rna: bool,
 }
 
 #[pymethods]
@@ -33,13 +20,34 @@ impl ViennaRNA {
 
     #[new]
     #[pyo3(signature = (
+        params = "rna_default",
         celsius=37.0,
     ))]
  
-    fn new(celsius: f64) -> Self {
-        Self { 
-            model: VRNA::from_thermo_params(&RNA_TURNER_2004, celsius)
-        }
+    fn new(params: &str, celsius: f64) -> PyResult<Self> {
+        let mut is_rna = true;
+        let thermo = match params {
+            "rna_default" => &RNA_TURNER_2004,
+            "rna_extended" => &RNA_EXTENDED,
+            "dna" => {
+                is_rna = false;
+                &DNA_MATHEWS_2004
+            },
+            _ => {
+                return Err(PyValueError::new_err(
+                    format!(
+                        "Unknown parameter set '{}'. \
+                         Valid options are: 'rna_default', 'rna_extended', 'dna'.",
+                        params
+                    )
+                ));
+            }
+        };
+
+        Ok(Self { 
+            model: VRNA::from_thermo_params(thermo, celsius),
+            is_rna,
+        })
     }
 
     fn energy_of_structure(
@@ -48,8 +56,12 @@ impl ViennaRNA {
         structure: &str,
     ) -> PyResult<f64> {
 
-        let sequence = NucleotideVec::try_from(sequence)
-            .map_err(|e| PyValueError::new_err(e.to_string()))?;
+        let sequence = match self.is_rna {
+            true => NucleotideVec::try_from_rna(sequence)
+                .map_err(|e| PyValueError::new_err(e.to_string()))?,
+            false => NucleotideVec::try_from_dna(sequence)
+                .map_err(|e| PyValueError::new_err(e.to_string()))?,
+        };
 
         let pairings = MultiPairTable::try_from(structure)
             .map_err(|e| PyValueError::new_err(e.to_string()))?;