+++ crates/fuzzyfold/CHANGELOG.md
@@ -2,9 +2,14 @@
 
 All notable changes to this crate will be documented in this file.
 
-## development
-- preliminary ff-tledit interface
-- changed ff-timecourse output to *.nxy
+## [0.4.2] - 2026-02-26
+## Added
+- fasta flag for ff-trajectory
+- title option for ff-timecourse
+
+## Changed
+- ff-timecourse output format is now *.nxy
+- Now allowing negative temperature (Celsius).
 
 ## [0.4.1] - 2026-02-26
 ## Changed